+++ src/main/java/frc/robot/config/game/reefscape2025/RobotConfigStub.java
@@ -21,7 +21,10 @@ public class RobotConfigStub extends RobotConfig {
       drive.setPose(new Pose2d(new Translation2d(1, 1), new Rotation2d()));
     }
 
-    /* We create a GizmoSubsystem using the MotorIOStub implementation of a MotorIO.  We set inverted=true because we want positive voltage to result in spinning clockwise.  The inversion depends on how the motor is physically connected to the gizmo, so specific to this robot configuration. */
+    // We create a GizmoSubsystem using the MotorIOStub implementation of a MotorIO.  We set
+    // inverted=true because we want positive voltage to result in spinning clockwise. The inversion
+    // depends on how the motor is physically connected to the gizmo, so specific to this robot
+    // configuration.
     gizmo = new GizmoSubsystem(new MotorIOStub(true));
   }
